코드 간소화
